#17077. Update PaymentOut to use Money instead of BigDecimal
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@24092 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-services/src/main/java/ish/oncourse/services/paymentexpress/PaymentOutSupport.java
@@ -38,7 +38,7 @@ public class PaymentOutSupport implements IPaymentSupport<PaymentOut, PaymentOut
 
         StringBuilder transactionDetails = new StringBuilder("Preparing payment transaction data. ");
 
-        details.setAmount(PaymentExpressUtil.translateInputAmountAsDecimalString(paymentOut.getTotalAmount()));
+        details.setAmount(PaymentExpressUtil.translateInputAmountAsDecimalString(paymentOut.getTotalAmount().toBigDecimal()));
         transactionDetails.append("amount: ").append(details.getAmount());
 
         details.setDpsTxnRef(paymentOut.getPaymentInTxnReference());
